+++ SustitucionMOAModel/Entities/Aprobaciones.cs
@@ -62,7 +62,7 @@ namespace SustitucionMOAModel.Entities
 
         public string Referencia { get; set; }
 
-        public int? Cantidad { get; set; }
+        public double? Cantidad { get; set; }
 
         public string Descripcion_ES { get; set; }
